requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
environs==9.5.0
marshmallow==3.14.1
Pillow==9.0.1
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
